clean up, fix test
