+++ dymos/__init__.py
@@ -5,5 +5,4 @@ from .transcriptions import GaussLobatto, Radau, ExplicitShooting, Analytic
 from .transcriptions.grid_data import GaussLobattoGrid, RadauGrid, UniformGrid
 from .trajectory.trajectory import Trajectory
 from .run_problem import run_problem
-from .load_case import load_case
 from ._options import options
